+++ CHANGELOG.md
@@ -9,6 +9,9 @@
   - Sheets that fail checks now have an alert next to them on the sheets index
   - An alert message can be added to a check that describes why the sheet is
     flagged on the sheets index
+  - Check names can now be used to filter sheets on the sheet index:
+    - `checks:any`
+    - `checks:NAME`
 - **Event Changes**
   - Event slugs now autofill when writing a new event name
   - Events can be archived from the event index